moved detailview to separate file
